Chores: cleanup dependencies and isolate macro subcrate
- Move dependencies that used only in tests to [dev-dependencies]
- isolate the `make_model` proc-macros to the subcrate, move its usage
to the main crate (`cif_chomper`).
