+++ Cargo.toml
@@ -1,7 +1,11 @@
 [workspace]
 resolver = "3"
-members = ["cif_chomper_core", "cif_chomper_macros", "cif_chomper", "cif_chomper_macros_core"]
+members = [
+    "cif_chomper_core", 
+    "cif_chomper_macros", 
+    "cif_chomper",
+]
 
 [workspace.dependencies]
 log = "0.4.28"
-syn = {version = "2.0.106", features = ["full"]}
+syn = "2.0.108"
