+++ README.md
@@ -4,3 +4,11 @@
 
 A nom-based CIF parser with (forthcoming) automatic generation of structs from
 the CIF dictionary definition.
+
+## Development
+
+To compile locally, need `cif_core` to include the full specification.
+
+```console
+git clone --recurse-submodules https://github.com/d-perl/cif_chomper
+```
