+++ cif_chomper_macros/Cargo.toml
@@ -1,13 +1,15 @@
-[lib]
-proc-macro = true
-
 [package]
 name = "cif_chomper_macros"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-cif_chomper_macros_core = {path = "../cif_chomper_macros_core"}
-cif_chomper_core = {path = "../cif_chomper_core"}
 syn = {workspace = true}
+proc-macro2 = "1.0.101"
+quote = "1.0.41"
 
+[dev-dependencies]
+trybuild = "1.0.113"
+
+[lib]
+proc-macro = true
