--- cif_chomper_macros_core/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "cif_chomper_macros_core"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-cif_chomper_core = {path = "../cif_chomper_core"}
-syn = {workspace = true}
-proc-macro2 = "1.0.101"
-quote = "1.0.41"
